Reduced Code Density
Applied appropriate Jinja Templating to reduce code density
